add review is ready in single course

+++ database/migrations/2024_03_15_214249_create_course_reviews_table.php
@@ -22,7 +22,7 @@ return new class extends Migration
             $table->string('title');
             $table->text('message');
             $table->boolean('status')->default(false);
-            $table->unsignedTinyInteger('rating')->default(1);
+            $table->decimal('rating')->default(1);
             $table->timestamps();
         });
     }
